Global user create and edit forms responsive and updated
